Moved the program cards a little bit
