+++ src/main/java/inf112/roborally/app/tile/RotateRight.java
@@ -18,6 +18,6 @@ public class RotateRight extends AbstractFunctionTile {
 
     @Override
     public void execute(Robot player) {
-        player.rotate(270);
+        player.rotate(-90);
     }
 }
